refactor(vc): extract vc_resolution.go from vc_service.go
Pure move of read-path query helpers into vc_resolution.go. Same
services package, no behavior change.

Moves: QueryExecutionVCs, GetExecutionVCByExecutionID, ListWorkflowVCs,
ListAgentTagVCs.

Refs #415
