+++ control-plane/internal/services/vc_service.go
@@ -342,38 +342,6 @@ func (s *VCService) generateVCID() string {
 	return fmt.Sprintf("vc-%s", uuid.New().String())
 }
 
-// QueryExecutionVCs queries execution VCs with filters.
-func (s *VCService) QueryExecutionVCs(filters *types.VCFilters) ([]types.ExecutionVC, error) {
-	if !s.config.Enabled {
-		return nil, fmt.Errorf("DID system is disabled")
-	}
-	return s.vcStorage.QueryExecutionVCs(filters)
-}
-
-// GetExecutionVCByExecutionID retrieves a single execution VC by execution identifier.
-func (s *VCService) GetExecutionVCByExecutionID(executionID string) (*types.ExecutionVC, error) {
-	if !s.config.Enabled {
-		return nil, fmt.Errorf("DID system is disabled")
-	}
-	return s.vcStorage.GetExecutionVCByExecutionID(executionID)
-}
-
-// ListWorkflowVCs lists all workflow VCs.
-func (s *VCService) ListWorkflowVCs() ([]*types.WorkflowVC, error) {
-	if !s.config.Enabled {
-		return nil, fmt.Errorf("DID system is disabled")
-	}
-	return s.vcStorage.ListWorkflowVCs()
-}
-
-// ListAgentTagVCs returns all non-revoked agent tag VCs.
-func (s *VCService) ListAgentTagVCs() ([]*types.AgentTagVCRecord, error) {
-	if !s.config.Enabled {
-		return nil, fmt.Errorf("DID system is disabled")
-	}
-	return s.vcStorage.ListAgentTagVCs(context.Background())
-}
-
 // collectDIDResolutionBundle collects all unique DIDs from the VC chain and resolves their public keys.
 func (s *VCService) collectDIDResolutionBundle(executionVCs []types.ExecutionVC, workflowVC *types.WorkflowVC) (map[string]types.DIDResolutionEntry, error) {
 	bundle := make(map[string]types.DIDResolutionEntry)
